pra 98 fix: обновление из ветки спринт_6
