+++ packages/client/src/pages/Forum/components/CommentsPage/CommentsPage.module.scss
@@ -3,5 +3,6 @@
 .block {
   display: flex;
   flex-flow: column;
+  margin-top: 30px;
   height: 500px;
 }
