+++ packages/client/src/pages/Forum/part/ButtonCreateTopic/ButtonCreateTopic.tsx
@@ -1,21 +1,21 @@
-import React, { useCallback } from 'react';
-import { useNavigate } from 'react-router-dom';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
 import { Button } from 'src/components/Button';
+import { useNavigator } from 'src/hooks/useNavigator';
 import styles from './ButtonCreateTopic.module.scss';
 
 export const ButtonCreateTopic = () => {
-    const navigate = useNavigate();
+    const { t } = useTranslation();
+    const navigator = useNavigator();
 
-    const handleCreateTopic = useCallback(() => {
-        navigate('/forum/createtopic');
-    }, [navigate]);
+    const handleCreateTopic = () => navigator('createtopic');
 
     return (
         <Button
             regular
             className={styles.button_create_topic}
             onClick={handleCreateTopic}>
-            POST NEW TOPIC
+            {t('postNewTopic')}
         </Button>
     );
 };
