+++ packages/client/src/pages/Forum/part/CommentList/CommentList.module.scss
@@ -13,7 +13,7 @@
     flex-flow: column;
     gap: 10px;
     overflow: auto;
-    height: 150px;
+    height: 250px;
     @include scroll($bg-color-main, $main_block_bg_color, $bg-color-secondary);
   }
