+++ packages/client/src/pages/Forum/part/CommentList/CommentList.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { useTranslation } from 'react-i18next';
 import { ErrorBoundary } from 'src/components/ErrorBoundary';
 import { Comment } from '../Comment';
 import { IComment } from '../Comment/types';
@@ -6,9 +7,11 @@ import styles from './CommentList.module.scss';
 import { Props } from './types';
 
 export const CommentList = ({ comments }: Props) => {
+    const { t } = useTranslation();
+
     return (
         <ErrorBoundary>
-            <div className={styles.title}>COMMENTS</div>
+            <div className={styles.title}>{t('comments')}</div>
             <div className={styles.list}>
                 {comments.map((comment: IComment, index: number) => {
                     return <Comment key={index} comment={comment} />;
